refactor: 상품 상세 페이지 코드 구조 수정
